reporting_obligations/requirements.txt
# uvicorn==0.12.2
fastapi==0.61.2
python-multipart==0.0.5
requests==2.25.0
dkpro-cassis==0.4.0
rdflib==5.0.0
SPARQLWrapper==1.8.5